requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.1
apache-airflow-providers-common-sql==1.6.1
apache-airflow-providers-ftp==3.4.2
apache-airflow-providers-google==10.5.0
apache-airflow-providers-http==4.5.0
apache-airflow-providers-imap==3.2.2
apache-airflow-providers-sqlite==3.4.3
apache-airflow==2.6.3
apispec==5.2.2
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backoff==2.2.1
bandit==1.7.5
blinker==1.6.2
cachelib==0.9.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2022.9.24
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
charset-normalizer==3.2.0
click==8.1.3
click==8.1.6
clickclick==20.10.2
colorama==0.4.5
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.3
db-dtypes==1.1.1
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.1
docutils==0.20.1
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
Flask-AppBuilder==4.3.1
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
Flask==2.2.5
frozenlist==1.4.0
gcloud-aio-auth==4.2.3
gcloud-aio-bigquery==6.3.0
gcloud-aio-storage==8.3.0
gitdb==4.0.10
GitPython==3.1.32
google-ads==21.2.0
google-api-core==2.11.0
google-api-python-client==2.95.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.22.0
google-cloud-aiplatform==1.28.1
google-cloud-appengine-logging==1.3.1
google-cloud-audit-log==0.2.5
google-cloud-automl==2.11.2
google-cloud-bigquery-datatransfer==3.12.0
google-cloud-bigquery-storage==2.22.0
google-cloud-bigquery==3.11.4
google-cloud-bigtable==2.20.0
google-cloud-build==3.19.0
google-cloud-compute==1.13.0
google-cloud-container==2.28.0
google-cloud-core==2.3.3
google-cloud-datacatalog==3.14.0
google-cloud-dataflow-client==0.8.4
google-cloud-dataform==0.5.2
google-cloud-dataplex==1.6.0
google-cloud-dataproc-metastore==1.12.0
google-cloud-dataproc==5.4.2
google-cloud-dlp==3.12.2
google-cloud-kms==2.19.0
google-cloud-language==2.10.1
google-cloud-logging==3.6.0
google-cloud-memcache==1.7.2
google-cloud-monitoring==2.15.1
google-cloud-orchestration-airflow==1.9.1
google-cloud-os-login==2.10.0
google-cloud-pubsub==2.18.1
google-cloud-redis==2.13.1
google-cloud-resource-manager==1.10.2
google-cloud-secret-manager==2.16.2
google-cloud-spanner==3.38.0
google-cloud-speech==2.21.0
google-cloud-storage-transfer==1.9.1
google-cloud-storage==2.10.0
google-cloud-tasks==2.14.0
google-cloud-texttospeech==2.14.1
google-cloud-translate==3.11.3
google-cloud-videointelligence==2.11.3
google-cloud-vision==3.4.4
google-cloud-workflows==1.10.2
google-crc32c==1.5.0
google-re2==1.1
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
graphviz==0.20.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio-gcp==0.2.2
grpcio-status==1.56.2
grpcio==1.56.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
idna==3.4
importlib-resources==6.0.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
json-merge-patch==0.2
jsonschema-specifications==2023.7.1
jsonschema==4.18.6
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
looker-sdk==23.12.0
Mako==1.2.4
markdown-it-py==3.0.0
Markdown==3.4.4
MarkupSafe==2.1.3
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
marshmallow==3.20.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.23.3
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.0.10
ordered-set==4.1.0
packaging==23.1
pandas-gbq==0.19.2
pandas==1.5.0
pandas==2.0.3
pathspec==0.9.0
pbr==5.11.1
pendulum==2.1.2
pluggy==1.2.0
prison==0.2.1
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.12
pydata-google-auth==1.8.2
Pygments==2.15.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2022.2.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.30.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
requests==2.28.1
requests==2.31.0
rfc3339-validator==0.1.4
rich-argparse==1.2.0
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
setproctitle==1.3.2
Shapely==1.8.5.post1
six==1.16.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sqlalchemy-bigquery==1.7.0
SQLAlchemy-JSONField==1.0.1.post0
sqlalchemy-spanner==1.6.2
SQLAlchemy-Utils==0.41.1
SQLAlchemy==1.4.49
sqlparse==0.4.4
stevedore==5.1.0
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
typer==0.6.1
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.12
urllib3==1.26.16
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
XlsxWriter==3.0.3
yarl==1.9.2